fix response to a GET with Range request header specifying multiple ranges
Motivation:

A GET request may include the Range request header, indicating that only
part of the entity is desired by the client.  Byte range, selecting the
content by byte offset, is a well-defined and commonly supported option.

Responses for requests with a Range header containing a single
byte-range differ from those for multiple byte-ranges.  In the former,
the output is returned as the response entity while in the latter the
multiple byte-ranges are combined with additional metadata in the
entity.

Currently the output for multiple byte-range GET requests is badly
formed.  It contains (at most) a single byte and the additional metadata
headers are badly separated.

Modification:

Fix range handling: add missing currentRange assignment and avoid
assumption that ranges are in-order.

Fix multi-range headers: adding missing new-line characters.

Result:

Support GET requests with a Range header specifying multiple byte
ranges.