Fix race-condition on PUT auto-creating collections.
Motivation:

There is a race-condition with multiple requests that target the same
missing collection.  These concurrent requests will all attempt to
create the same collection, resulting in one succeeding and the
remaining requests failing to create the collection.

The current PutHandler does not expect this failure, with the result
that PUT requests will fail.

Modification:

Update PUT request code that creates missing parent collections so that
it handles such failures.  If it receives a BadRequestException then it
simply retries fetching the child.

Result:

Multiple concurrent requests that target the same missing directory now
succeed.